Add features:
Use KaTeX for math typsetting.
When we are in Home or the page variable `katex` is true, KaTeX
is loaded.

https://katex.org/docs/autorender.html

Some variables are now optional.  Because some conferences are held
on WEB, some variables (e.g., institute, city, country) are no
longer used.

New variables (note/file) are introduced.  Use them for posters for
example.